refactor + docs
