+++ main-app/src/Components/Infobox/Infobox.js
@@ -1,0 +1,1 @@
+// TODO: für was wollte ich das brauchen?
