+++ main-app/src/Components/MainSites/Footer/Cookies.js
@@ -1,6 +1,11 @@
 import React from "react";
 
 
+/**
+ * Cookie Policy. Not implemented yet. At the moment only general Information.
+ * @returns {Element}
+ * @constructor
+ */
 const Cookies = () =>
 {
