+++ main-app/src/Components/MainSites/Footer/Credits.js
@@ -1,6 +1,11 @@
 import React from "react";
 
 
+/**
+ * Simple credits, implemented when worked as a team.
+ * @returns {Element}
+ * @constructor
+ */
 const Credits = () =>
 {
 
@@ -10,7 +15,7 @@ const Credits = () =>
             <h1 className={"mt-20 italic"}>"I am not young enough to know everything."</h1>
             <h2 className={"mt-5"}>Oscar Wilde</h2>
             <br/>
-            <p className={"mt-20"}>Quelle: <a href={"https://beruhmte-zitate.de/zitate/1982587-oscar-wilde-ich-bin-nicht-jung-genug-um-alles-zu-wissen/"} target="_blank">Link...</a> </p>
+            <p className={"mt-20"}>Source: <a href={"https://beruhmte-zitate.de/zitate/1982587-oscar-wilde-ich-bin-nicht-jung-genug-um-alles-zu-wissen/"} target="_blank">Link...</a> </p>
 
         </div>
     )
