+++ main-app/src/Components/MainSites/MainMenuHeader/Home.js
@@ -3,12 +3,21 @@ import Tile from "../../Tile/Tile";
 
 //import "../../Styles/MainSites/Grid_home.css"
 
+/**
+ * Landing page of the website(s).
+ * Content, without header and footer.
+ * Contains a hierarchical structures of tiles which serve navigation to different topics.
+ * will be defined in a JSON file or directory-structure.
+ *
+ * @returns {Element}
+ * @constructor
+ */
 const Home = () => {
 
     const tileData = [
         {
             title: "Infos",
-            subtitle: "Infocenter for trainings",
+            subtitle: "Info-Center for trainings",
             cornerImageUrl: "../../Images/Logos/Infos.jpg",
             linkTo: "/infos",
         },
