+++ main-app/src/Components/MnemoMaxTrainer/ComputationsDataGet.js
